+++ src/main/kotlin/app/cash/licensee/pluginExtension.kt
@@ -22,6 +22,7 @@ import org.gradle.api.Action
 import org.gradle.api.Named
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.artifacts.Dependency
+import org.gradle.api.artifacts.ExternalModuleDependency
 import org.gradle.api.provider.MapProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
@@ -129,6 +130,19 @@ interface LicenseeExtension {
     dependencyProvider: Provider<out Dependency>,
   ) {
     allowDependency(dependencyProvider, {})
+  }
+
+  fun allowDependency(
+    externalModuleDependency: ExternalModuleDependency,
+    options: Action<AllowDependencyOptions> = Action { },
+  )
+
+  /** @suppress */
+  @JvmSynthetic
+  fun allowDependency(
+    externalModuleDependency: ExternalModuleDependency,
+  ) {
+    allowDependency(externalModuleDependency, {})
   }
 
   /**
@@ -358,12 +372,27 @@ internal abstract class MutableLicenseeExtension : LicenseeExtension {
       dependencyProvider.map {
         mapOf(
           DependencyCoordinates(
-            group = requireNotNull(it.group) { "group was null in allowDependency for ${it.name}" },
+            group = it.getLicenseeGroup(),
             artifact = it.name,
-            version = requireNotNull(it.version) { "version was null in allowDependency for ${it.name}" },
+            version = it.getLicenseeVersion(),
           ) to Optional.ofNullable(optionsImpl.setReason),
         )
       },
+    )
+  }
+
+  private fun Dependency.getLicenseeGroup() = requireNotNull(group) { "group was null in allowDependency for $name" }
+  private fun Dependency.getLicenseeVersion() = requireNotNull(version) { "version was null in allowDependency for $name" }
+
+  override fun allowDependency(
+    externalModuleDependency: ExternalModuleDependency,
+    options: Action<AllowDependencyOptions>,
+  ) {
+    allowDependency(
+      groupId = externalModuleDependency.getLicenseeGroup(),
+      artifactId = externalModuleDependency.name,
+      version = externalModuleDependency.getLicenseeVersion(),
+      options = options,
     )
   }